(ref) put createRequests and createDrivers functions into the helpers file
