+++ service/src/database/setup/seed.js
@@ -1,54 +1,10 @@
-/* eslint no-plusplus: ["error", { "allowForLoopAfterthoughts": true }] */
-
-import faker from 'faker';
-
 import db from '../index';
 import helpers from './helpers';
 
-const { firstName, lastName } = faker.name;
+const { pgKnex } = db;
 
-const { pgKnex, st } = db;
+const { createDrivers, createRequests } = helpers;
 
-const { createLocation, createTime } = helpers;
-
-const createDrivers = (count) => {
-  const drivers = [];
-
-  for (let i = 0; i < count; i++) {
-    const createdAt = createTime();
-    const updatedAt = createTime(createdAt.getTime());
-    const location = createLocation();
-
-    const info = {
-      first_name: firstName(),
-      last_name: lastName(),
-      joined: createdAt.toISOString(),
-      last_checkin: updatedAt.toISOString(),
-      available: true,
-      booked: false,
-      location: st.geomFromText(location, 4326),
-    };
-    drivers.push(info);
-  }
-
-  return drivers;
-};
-
-const createRequests = (count) => {
-  const requests = [];
-
-  for (let i = 0; i < count; i++) {
-    const location = createLocation();
-
-    const info = {
-      start_loc: st.geomFromText(location, 4326),
-    };
-
-    requests.push(info);
-  }
-
-  return requests;
-};
 
 const start = new Date();
 console.log('start: ', start.toISOString());
